Moved properties to their own file, small codestyle fixes
